requirements.txt
numpy==1.17.0
Pillow==6.1.0
torch==1.2.0
torchvision==0.4.0